+++ R/chat-tools.R
@@ -10,10 +10,17 @@ match_tools <- function(turn, tools) {
   }
 
   turn@contents <- map(turn@contents, function(content) {
-    if (!is_tool_request(content)) {
+    if (is_tool_request(content)) {
+      content@tool <- content@tool %||% tools[[content@name]]
       return(content)
     }
-    content@tool <- tools[[content@name]]
+
+    if (is_tool_result(content)) {
+      content@request@tool <-
+        content@request@tool %||% tools[[content@request@name]]
+      return(content)
+    }
+
     content
   })