+++ R/provider-ollama.R
@@ -38,7 +38,7 @@
 #' }
 chat_ollama <- function(
   system_prompt = NULL,
-  base_url = "http://localhost:11434",
+  base_url = Sys.getenv("OLLAMA_BASE_URL", "http://localhost:11434"),
   model,
   seed = NULL,
   api_args = list(),