+++ R/provider-snowflake.R
@@ -284,7 +284,6 @@ method(as_json, list(ProviderSnowflakeCortex, Turn)) <- function(provider, x) {
   }
   list(
     role = x@role,
-    content = content,
     content_list = as_json(provider, x@contents)
   )
 }
@@ -435,6 +434,10 @@ snowflake_keypair_token <- function(
     fp <- openssl::base64_encode(
       openssl::sha256(openssl::write_der(key$pubkey))
     )
+    if (grepl(".+\\.privatelink$", account)) {
+      # account identifier is everything up to the first period
+      account <- gsub("^([^.]*).+", "\\1", account)
+    }
     sub <- toupper(paste0(account, ".", user))
     iss <- paste0(sub, ".SHA256:", fp)
     # Note: Snowflake employs a malformed issuer claim, so we have to inject it