+++ NEWS.md
@@ -1,3 +1,5 @@
+# ellmer (development version)
+
 # ellmer 0.3.0
 
 ## New features