+++ NEWS.md
@@ -1,5 +1,7 @@
 # ellmer (development version)
 
+* In the `value_turn()` method for OpenAI providers, `usage` is checked if `NULL` before logging tokens to avoid errors when streaming with some OpenAI-compatible services (#706, @stevegbrooks).
+
 * `chat_anthropic()` drops empty assistant turns to avoid API errors (#710).
 
 * `contents_record()` and `contents_replay()` now record and replay custom classes that extend ellmer's `Turn` or `Content` classes (#689).